core: Extracted test helper for `RetryAfterExtension`.
This helper class is useful for all kind of extension actors
implementing some retry logic. Therefore, it makes sense to have it
available in a reusable form. It is going to be reused in a future
commit.
